% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all.R
\name{allyears}
\alias{allyears}
\title{All Years}
\usage{
allyears()
}
\value{
The complete data frame with all the information from the Williams College Course Catalog,
the same as the allyrs dataset
}
\description{
allyrs runs getyr for all years and combines them into a single data frame
}
