% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_strings.R
\name{clean1}
\alias{clean1}
\title{Clean Data Part One}
\usage{
clean1("file")
}
\arguments{
\item{file}{A file that is the output of makestrings(), a set of complete strings,
each a person in the Williams College Course Catalog.}
}
\value{
clean1 returns partially cleaned data ready for clean2
}
\description{
clean1() is the first part of the data cleaning series
}

