% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_strings.R
\name{clean3}
\alias{clean3}
\title{Clean Data Part Three}
\usage{
clean3("00") gives partially cleaned entries from the Williams College Course Catalog.
}
\arguments{
\item{file}{A file that is the output of clean2(), a set of partially cleaned strings,
each a person in the Williams College Course Catalog.}

\item{year}{A year}
}
\value{
clean3 returns partially cleaned data ready for clean2
}
\description{
clean3() is the third part of the data cleaning series, it focuses on departmental honors
}

