% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_strings.R
\name{clean4}
\alias{clean4}
\title{Clean Data Part Four}
\usage{
clean4("file", "year")
}
\arguments{
\item{file}{A file that is the output of clean3(), a set of partially cleaned strings,
each a person in the Williams College Course Catalog.}

\item{year}{The year after the course catalog came out, for the 1999-2000 catalog, the year would be "00", the same as the file}
}
\value{
clean4 returns completely cleaned dataset for a single year.
}
\description{
clean4() is the fourth part of the data cleaning series, it focuses on adding gender information.
}
