% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_data.R
\name{combine_data}
\alias{combine_data}
\title{combine_data}
\usage{
combine_data()
}
\value{
The output of combine_data() is a data frame
}
\description{
combine_data() takes multiple Williams College Catalog data frames and
puts them together to form a single, clean data frame
}

