% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answers2.R
\name{dept_honors}
\alias{dept_honors}
\title{Departmental Honors}
\usage{
dept_honors()
}
\value{
A bar graph showing the total number of graduates that earned some level of departmental honors, by major and gender.
}
\description{
dept_honors shows the total number of graduates that earned some level of departmental honors, by major and gender.
}
