% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_strings.R
\name{findstrings}
\alias{findstrings}
\title{Find Strings}
\usage{
findstrings("00") cuts the 2000 catalog to just the graduated seniors mentioned in the document
}
\arguments{
\item{file}{The file, a Williams College Course Catalog, in .txt form, to be cut down to just the list of graduated seniors.}
}
\value{
findstrings(file) returns a tbl_df() of the section of the Williams College Course Catalog that contains the list of graduated seniors.
}
\description{
findstrings() cuts the txt version of the Williams College Course Catalog to isolate the section
 of the Catalog that is just list of graduating seniors.
}

