% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{get_data}
\usage{
get_data(file)
}
\arguments{
\item{file}{The file, a Williams College catalog, to be converted from pdf and cleaned}
}
\value{
The output of get_data() is a data frame
}
\description{
get_data() converts the pdf Williams College Catalog to a partially clean data frame
}

