% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all.R
\name{getyr}
\alias{getyr}
\title{Get Year}
\usage{
getyr("file")
}
\arguments{
\item{file}{The .pdf file to be converted to a data frame}
}
\value{
A complete data frame for a single year
}
\description{
getyr combines all of the necesary functions and gives a complete data frame for a single year
}

