% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answers.R
\name{gradcount}
\alias{gradcount}
\title{Graduate Count}
\usage{
gradcount()
}
\value{
A bar graph of graduating seniors by year
}
\description{
gradcount gets the count of graduates each year and graphs the information
}

