% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initials.R
\name{initials}
\alias{initials}
\title{visual}
\usage{
visual() shows...
}
\value{
A graph showing the joint and marginal distribution of initials
}
\description{
visual() makes a graphic showing the frequency of two variables and their marginal distributions.
It is intended to be used to show distribution of initials
}

