% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latin_honors.R
\name{latin_honors}
\alias{latin_honors}
\title{Latin Honors}
\usage{
latin_honors()
}
\value{
A bar graph showing the percentage of each graduating class awarded each level of latin honors.
}
\description{
latin_honors shows the percentage of each graduating class that is awarded latin honors.
}
