% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answers2.R
\name{latin_honors_gender}
\alias{latin_honors_gender}
\title{Latin Honors and Gender}
\usage{
latin_honors_gender()
}
\value{
A bar graph showing the total number of students of each predicted gender
 who received each level of latin honors.
}
\description{
latin_honors_gender shows the total number of students of each predicted gender
 who received each level of latin honors.
}
