% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answers.R
\name{namelength}
\alias{namelength}
\title{Name Length}
\usage{
namelength()
}
\value{
A data frame of all graduates with their name, the number of characters and their graduation year
}
\description{
namelength counts the number of letters in each name
}
