% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answers2.R
\name{pbk}
\alias{pbk}
\title{Phi Beta Kappa}
\usage{
pbk()
}
\value{
A bar graph showing the percentage of each graduating class awarded Phi Beta Kappa
}
\description{
pbk shows the percentage of each graduating class that is awarded Phi Beta Kappa.
}
