% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latin_honors.R
\name{total_latin_honors}
\alias{total_latin_honors}
\title{Total Latin Honors}
\usage{
total_latin_honors()
}
\value{
A bar graph showing the percentage of each graduating class attaining at least each level of latin honors.
There are also black lines that indicate the guidelines given in the Williams College Course Catalog.
}
\description{
total_latin_honors shows cumulative latin honors for comparison with guidelines given by the Williams College Faculty.
}

