% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_strings.R
\name{clean2}
\alias{clean2}
\title{Clean Data Part Two}
\usage{
clean2("00") gives partially cleaned entries from the Williams College Course Catalog.
}
\arguments{
\item{file}{A file that is the output of clean1(), a set of partially cleaned strings,
each a person in the Williams College Course Catalog.}
}
\value{
clean2 returns partially cleaned data ready for clean3
}
\description{
clean2() is the second part of the data cleaning series, it focuses on correcting last names that have spaces
and have been read incorrectly as a result.
}

