% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_strings.R
\name{clean3}
\alias{clean3}
\title{Clean Data Part Three}
\usage{
clean3("file", "year")
}
\arguments{
\item{file}{A file that is the output of clean2(), a set of partially cleaned strings,
each a person in the Williams College Course Catalog.}

\item{year}{The year after the course catalog came out, for the 1999-2000 catalog, the year would be "00", the same as the file.}
}
\value{
clean3 returns partially cleaned data ready for clean4
}
\description{
clean3() is the third part of the data cleaning series, it focuses on departmental honors.
}
