% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initials.R
\name{initials}
\alias{initials}
\title{Initials}
\usage{
initials()
}
\value{
A graph showing the joint and marginal distribution of initials
}
\description{
initials() makes a graphic showing the frequency of initials and the marginal distributions of first and last initials.
}

