% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_strings.R
\name{makestrings}
\alias{makestrings}
\title{Find Strings}
\usage{
makestrings("00") takes the output of findstrings("00") and gives complete lines from the Williams College Course Catalog.
}
\arguments{
\item{file}{The file, a section Williams College Course Catalog, in .txt form. This section may not be complete lines.}
}
\value{
The output of makestrings is a tbl_df() of complete lines from the graduated seniors section of the Williams College Course Catalog.
}
\description{
makestrings() takes the list of names given by findstrings() and combines
incomplete entries to form complete entries of graduating seniors
}

