% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_strings.R
\name{pdf2txt}
\alias{pdf2txt}
\title{PDF to Text}
\usage{
pdf2txt("00") converts the pdf from 1999-2000 to txt
}
\arguments{
\item{file}{The file, a Williams College Course Catalog, to be converted from pdf}
}
\value{
The output of pdftotxt is a txt version of the Williams College Course Catalog
in the same location as the .pdf that it took as input.
}
\description{
pdf2txt() converts Williams College Course Catalogs from .pdf to .txt
}

