% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcgrads.R
\docType{package}
\name{wcgrads}
\alias{wcgrads}
\alias{wcgrads-package}
\title{Williams College Graduates}
\description{
wcgrads cleans and examines data found in Williams College Course Catalogs about graduating seniors from 2000-2016.
}
\details{
The Williams College Course Catalog is published as a PDF document.
wcgrads converts that data into a readable dataframe.
The dataframe and sample analyses are then made available to package users.
}

